Create rock paper scissors
Create rock, paper and scissors game that allows a user to play the game against the computer.
The program prompts a user to enter a choice, writes to console
And compares to the computers choice.
It plays 5 rounds and adds up score for the winner after each rounds
The winner is logged at the end of the game to the console.